Get results from the Cristin API

get_cristin_results() takes in search parameters for the Cristin API and returns the results as a tibble

get_cristin_results(..., page = 1, per_page = 999, fields = "all", NVI = FALSE)

Arguments

...
Arguments to pass on to Cristin API. See 'API parameters' below for more information.

page
Page number for pagination purposes, set to 1 by default

per_page
Number of items per page, set to 999 by default

fields
'all' by default, if omitted will return a minimal set of fields

NVI
Logical determining whether to return only peer-reviewed publications or all results, set to 'FALSE' by default

Value

A tibble with results from the API call, with one row per result.

API parameters

The Cristin API recognises many search parameters. You can find the latest specification at the Cristin API documentation page.

Here are examples of the formatting of these parameters:

parameter	type	description

doi	string	DOI name (e.g. doi=10.1000/123456)

title The title of the result

contributor Author's name or Cristin person id

issn The issn of the result

unit Id (e.g. unit=185.53.18.10)

institution Id (one number, e.g. institution=185), name or acronym of the institution the contributors belong to

user A person's username in Cristin together with the institution id separated by ':' 'askeladd:185' means Cristin user 'askeladd' from the institution '185'

category Result category code, see Cristin category codes for specification

published_since Results published since and inclusive the given year, (yyyy), e.g: 2005

published_before Results published before and inclusive the given year, (yyyy), e.g: 2017

created_since Results created since and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 * created_before Results created before and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 * modified_since Results modified since and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17

modified_before Results modified before and inclusive the given date, (yyyy-mm-dd), e.g: 2005-03-17 * year_reported The year a result was reported * project_code Project code is the internal reference number used by funding source * funding_source Funding source code e.g: NFR * funding Funding source code e.g: NFR, and project_code together separated by ':' 'NFR:1234' means 'funding_source': 'NFR' with 'project_code': '1234' * lang Two letter language code that determines value of fields such as institution name, one of 'en' (default), 'nb', or 'nn'.

Examples

get_cristin_results(contributor = "25062")

#> # A tibble: 53 x 53
#>    cristin_result_~ result_url category_code title year_published
#>    <chr>            <chr>      <chr>         <chr> <chr>         
#>  1 235561           https://a~ ACADEMICLECT~ Rene~ 2008          
#>  2 243774           https://a~ ACADEMICLECT~ Cons~ 2009          
#>  3 249116           https://a~ ACADEMICLECT~ Wher~ 2009          
#>  4 254999           https://a~ ACADEMICLECT~ Unde~ 2010          
#>  5 255746           https://a~ ACADEMICLECT~ Enac~ 2010          
#>  6 255798           https://a~ ACADEMICLECT~ Bill~ 2010          
#>  7 251539           https://a~ ACADEMICLECT~ Rene~ 2010          
#>  8 1051805          https://a~ ACADEMICLECT~ Towa~ 2013          
#>  9 1051810          https://a~ LECTUREPOPUL~ Bør ~ 2013          
#> 10 1051802          https://a~ ACADEMICLECT~ The ~ 2013          
#> # ... with 43 more rows, and 48 more variables: funding_sources <list>,
#> #   open_access <chr>, original_language <chr>, organiser <chr>, links <list>,
#> #   year_reported <chr>, import_sources <list>, year_online <chr>,
#> #   volume <chr>, date_published <chr>, issue <chr>, number_of_pages <chr>,
#> #   projects <list>, international_standard_numbers <list>, year_printed <chr>,
#> #   category_name_en <chr>, contributors_url <chr>, contributors_count <int>,
#> #   contributors_preview <list>, created_date <chr>, last_modified_date <chr>,
#> #   event_name <chr>, event_location <chr>, event_date_from <chr>,
#> #   event_date_to <chr>, event_arranged_by_name <chr>, channel_title <chr>,
#> #   part_of_url <chr>, pages_from <chr>, pages_to <chr>, pages_count <chr>,
#> #   journal_cristin_journal_id <chr>, journal_name <chr>,
#> #   journal_international_standard_numbers <list>, journal_nvi_level <chr>,
#> #   journal_publisher_cristin_publisher_id <chr>, journal_publisher_name <chr>,
#> #   journal_publisher_url <chr>, journal_publisher_place <chr>,
#> #   publisher_name <chr>, publisher_cristin_publisher_id <chr>,
#> #   publisher_url <chr>, series_cristin_journal_id <chr>, series_name <chr>,
#> #   series_international_standard_numbers <list>,
#> #   classification_scientific_disciplines <list>,
#> #   classification_keywords <list>, summary <chr>